》完成用例新增及编辑，状态，删除联调 》完成项目编辑状态，状态，删除联调

+++ Mechanical_insect/app/admin/views.py
@@ -8,7 +8,7 @@
 
 from . import admin
 from flask import render_template, redirect, url_for, flash, session, request, abort
-from app.admin.forms import LoginForm, AdminForm, RoleForm, AuthFrom, ProjectFrom
+from app.admin.forms import LoginForm, AdminForm, RoleForm, AuthFrom, ProjectFrom, CaseFrom
 from app.models import Admin, User, Oplog, Adminlog, Userlog, Auth, Role, Project, Case
 from functools import wraps
 from app import db
@@ -444,7 +444,7 @@ def project_list(page=None):
 
 
 # 项目状态
-@admin.route("/project/status/<int:id><int:status>/", methods=["GET", "POST"])
+@admin.route("/project/status/<int:id>/<int:status>", methods=["GET", "POST"])
 @admin_login_req
 # @admin_auth
 def project_status(id=None, status=None):
@@ -454,13 +454,13 @@ def project_status(id=None, status=None):
     #     return redirect(url_for("admin.admin_list", page=1))
     if status == 0:
         project.status = 1
-        flash("停用角色成功！", "ok")
+        flash("停用项目成功！", "ok")
         oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="停用项目%s" % id)
         db.session.add(oplog)
         db.session.commit()
     else:
         project.status = 0
-        flash("启用角色成功！", "ok")
+        flash("启用项目成功！", "ok")
         oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="启用项目%s" % id)
         db.session.add(oplog)
         db.session.commit()
@@ -497,15 +497,42 @@ def project_add():
 
 
 # 项目删除
-@admin.route("/project/del/<int:id>/", methods=["GET"])
+@admin.route("/project/del/<int:id>/<int:status>", methods=["GET"])
 @admin_login_req
 # @admin_auth
-def project_del(id=None):
+def project_del(id=None, status=None):
     project = Project.query.get_or_404(int(id))
+    if status == 0:
+        flash('项目正在使用！', category='err')
+        return redirect(url_for('admin.project_list', page=1))
     db.session.delete(project)
     db.session.commit()
     flash("删除项目成功！", "ok")
     return redirect(url_for("admin.project_list", page=1))
+
+
+# 项目修改
+@admin.route("/project/edit/<int:id>/", methods=["GET", "POST"])
+@admin_login_req
+# @admin_auth
+def project_edit(id=None):
+    form = ProjectFrom()
+    project = Project.query.get_or_404(id)
+    if form.validate_on_submit():
+        data = form.data
+        project.version = data['version']
+        project.name = data["name"]
+        project.models = data["models"]
+        project.leader = data["leader"]
+        project.comment = data["comment"]
+        db.session.add(project)
+        db.session.commit()
+        flash("修改项目成功！", "ok")
+        oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="修改项目%s" % data['name'])
+        db.session.add(oplog)
+        db.session.commit()
+        return render_template("admin/project_edit.html", form=form, project=project)
+    return render_template("admin/project_edit.html", form=form, project=project)
 
 
 # 用例列表
@@ -518,3 +545,71 @@ def case_list(page=None):
     page_data = Case.query.join(User).filter(User.id == Case.user_id, ).order_by(Case.addtime.desc()).paginate(
         page=page, per_page=10)
     return render_template("admin/case_list.html", page_data=page_data)
+
+
+# 添加用例
+@admin.route("/case/add/", methods=["GET", "POST"])
+@admin_login_req
+# @admin_auth
+def case_add():
+    form = CaseFrom()
+    if form.validate_on_submit():
+        data = form.data
+        if Case.query.filter_by(name=data['name']).count() == 1:
+            flash('用例名称已存在！', category='err')
+            return redirect(url_for('admin.case_add'))
+        elif Case.query.filter_by(name=data['version']).count() == 1:
+            flash('用例编号已存在！', category='err')
+            return redirect(url_for('admin.case_add'))
+        case = Case(name=data["name"], version=data["version"], models=data["models"], user_id=session["admin"],
+                    case_leader=data["case_leader"], comment=data["comment"], Environment=data["Environment"],
+                    pass_num='',
+                    fail_num='', execute_count='', case_pass='', status=0)
+        db.session.add(case)
+        db.session.commit()
+        flash("添加用例成功！", "ok")
+        oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="添加用例：%s" % data['name'])
+        db.session.add(oplog)
+        db.session.commit()
+    return render_template("admin/case_add.html", form=form)
+
+
+# 用例状态
+@admin.route("/case/status/<int:id>/<int:status>", methods=["GET", "POST"])
+@admin_login_req
+# @admin_auth
+def case_status(id=None, status=None):
+    case = Case.query.get_or_404(id)
+    # if id == 1:
+    #     flash("无权停用，联系管理员！", "err")
+    #     return redirect(url_for("admin.admin_list", page=1))
+    if status == 0:
+        case.status = 1
+        flash("停用项目成功！", "ok")
+        oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="停用项目%s" % id)
+        db.session.add(oplog)
+        db.session.commit()
+    else:
+        case.status = 0
+        flash("启用项目成功！", "ok")
+        oplog = Oplog(admin_id=session["admin_id"], ip=request.remote_addr, reason="启用项目%s" % id)
+        db.session.add(oplog)
+        db.session.commit()
+    db.session.add(case)
+    db.session.commit()
+    return redirect(url_for("admin.case_list", page=1))
+
+
+# 用例删除
+@admin.route("/case/del/<int:id>/<int:status>", methods=["GET"])
+@admin_login_req
+# @admin_auth
+def case_del(id=None, status=None):
+    case = Case.query.get_or_404(int(id))
+    if status == 0:
+        flash('项目正在使用！', category='err')
+        return redirect(url_for('admin.case_list', page=1))
+    db.session.delete(case)
+    db.session.commit()
+    flash("删除项目成功！", "ok")
+    return redirect(url_for("admin.case_list", page=1))
